Move the common logic of View/Template to Core\BaseView
